Minor internal member list optimizations
• -_removeMemberWithNickname: now returns boolean whether user was
removed or not. -removeMember: uses this boolean to know whether to
post event to view controllers about the removal. No need to inform
them of a member being removed when they aren’t actually removed.
• -_removeMemberWithNickname: was also updated to call -removeObject:
on the length sorted member list instead of looking up the object’s
index on our own. I do not see a case where the user object would exist
in the length sorted array but not the primary so this change makes
sense at this time.