Add self.inUserInvoked*Request Helper Methods
This may not be the best or most pretty change, but it works. When an
self.inUserInvoked*Request property is set, a timer is enabled that
will store pointer to the property and reset it after 10 seconds if its
not already set back to NO by that point.

This change accounts for possible failures when the command related to
the property is not executed successfully and Textual is unable to
associate a particular error message with the property leaving the
property with a value of YES.