Improve tab completion logic
• Store selection range after tab completion is completed and compare
that against the selection range when user tab completes agains so we
know whether to start fresh.
• If we are completing nickname, user configured completion suffix is
empty, and we are not at the end of the line then we insert a
whitespace.
• If the user has selected text (selected text length > 0), then we use
that length as the completion suffix length and perform no other
matching.
• Added hidden preference to allow users who want the completion suffix
to cut everything until first space, to enable this behavior.