When WebView becomes hidden, cache scroll position
• When WebView becomes hidden (not selected), cache whether the WebView
was scrolled to the bottom at the time it became hidden.
• If the view was scrolled to the bottom when returning visibility to
WebView, pass parameter “wasViewingBottom” to
Textual.notifyDidBecomeVisible() to only scroll to adjust scroll
position if we were previously at bottom.
• TVCWebViewAutoScroll.m is asked for information about scroll position
to determine we are at the bottom to require passes at the DOM to
determine this information as its already tracked by the auto scroller.
• No longer call -notifyDidBecomeVisible when clearing a WebView. The
entire WebView is reconstructed when cleared so it makes little sense
to call this then. Reserve -notifyDidBecomeVisible for when actual
visibility (selection) changes occur.