Refactor chat filter help buttons
Using NSPopover proved to be quite unreliable for presenting help
information. It may have been a shotty implementation or failure with
the UI element in general, but its best to move away from it. Textual’s
knowledge base already contains help information for Chat Filter, so
we’ll just link to the relevant sections. This solution is good because
it allows errors to be corrected in real time as well.