Add +q to list of bad modes
IRCChannelMode excludes user modes defined in a network’s configuration
which typically includes +q for channel owner. As this mode was
typically excluded automatically, it was not manually labeled as a bad
mode.

However, some networks use +q to quiet users, which means that quieted
users would end up displaying in formatted output of IRCChannelMode
when they were not supposed to.

This commit labels +q as a mode to exclude from formatted output no
matter what as its a special mode.