Fix for highlight list dialog hanging
The Highlight List dialog performs complex math for each row added to
determine the height the row should be when the text is wrapped
relative to the column width. There is no good solution going back all
the way to Mountain Lion to do this with Auto Layout, so Textual does
it the hard way.

These calculations lagged the dialog to the point the main thread
became unresponsive. This change lazily defines the height of each row.

When a new row is added, the new default row height is returned (17).
Once the row is added, a block on the global queue (in the background)
determines the height, saves that value in a property, then informs the
table view that the height has changed.