Fix for scrolling becoming stuck
When TVCLogController is created or reset through the use of clearing
content, the property self.wasViewingBottomBeforeBecomingHidden is set
to YES which tells the view that next time its selected, it should
scroll all the way to the bottom showing the latest messages because
the user was probably already at the bottom before switching away.

When switching away, the value of this property is set by calling
-isViewingBottom in TVCWebViewAutoScroll. Turns out if the view is not
scrollable, -isViewingBottom will return NO when we would actually want
it to respond YES which means once the view fills up to the point it is
scrollable and the user switches back, Textual doesn’t automatically
scroll back to the bottom position because it believe its not supposed
to (maybe the user manually scrolled to a position).