JavaScript Everything
In anticipation of using WebKit2, all interactions with the DOM must be
performed through JavaScript.

This commit includes many things:
• TVCLogView received a very primitive compiler that can take in a
function name + arguments and output JavaScript. WK2 does not have a
way to invoke script objects so we must rely on evaluating blobs of
JavaScript on their own.
• TVCLogView also received many helper methods for interacting with the
DOM using this new compiler. Very strict type checking is used to make
sure the data we are getting back is the data that we want.
• Added autoScrollPrivate.js which is a drop in replacement for the
objective-c auto scroller that was being used (TVCWebViewAutoScroll)
• corePrivate.js received a significant overhaul. Everything that
existed in TVCLogController for interacting with the DOM using
Objective-C is now JavaScript.
• Added scriptSinkPrivate.js - This source file contains the
declaration of the “app” object for JavaScript ➜ Objective-C
communication. WK1 and WK2 interact with Objective-C using two
different interfaces. scriptSinkPrivate.js removes that complexity so
that style authors can focus on their style.

• And much more…