Handle dropped processes gracefully
When a Web Content process closes unexpectedly, the backing view is
reconstructed inside TVCLogController.

This commit changes the behavior of baseLayout.mustache. Previously,
during a clear, the value of #topic_bar was taken and set in the new
document as the token {formattedTopicValue}. If a view is lost, we will
not be able to get the #topic_bar value. To workaround this, the
{formattedTopicValue} token no longer exists. The topic is set using
Textual.setTopicBarValue() once the view is loaded.