Bump THOPluginProtocolCompatibilityMinimumVersion
For the first time since adding the constant,
THOPluginProtocolCompatibilityMinimumVersion has been bumped. Textual
6.0 contains significant changes in its underlying WebView
architecture. As there are a handful of plugins known to exist that
modify the WebView using certain structures that Textual exposes, it is
best not to load these plugins anymore to reduce crashes.

109c71f added version checking to THOPluginManager plus a whitelist of
bundle identifiers that are known not to touch the WebView and are
therefore safe.