Refactor build process
• Copy a separate Info.plist file depending on build destination (App
Store or not)
• Remove macros from GCC_PREPROCESSOR_DEFINITIONS. All macros are
already defaulted to 0 through TextualApplication.h - use
GCC_PREPROCESSOR_DEFINITIONS to enable, not disable. That doesn’t make
sense.
• Add comment so that those that modify GCC_PREPROCESSOR_DEFINITIONS
will understand that the macros defined by it only enable the code, not
resource files the code may be dependent on.
• Use “set -e” to stop scripts from executing if anything returns an
error.
• Run CleanUpResources.sh and UpdateVersionInfo.sh as pre-build scripts
to ensure all files are where they need to be before any building
occurs.
• Remove FindProvisioningProfile.sh - this file has not worked
correctly in many months.