Add TextualScroller.nextScrollTopValue
Do not move scroll horizon until a scroll event occurs. The next value
is stored in TextualScroller.nextScrollTopValue and is pushed to
TextualScroller.currentScrollTopValue when a scroll event occurs that
does not go above the current horizon.

This change is made to try to stop a user event occurring before
TextualScroller.currentScrollTopValue == TextualScroller.scrollHeight()