Refactor selection process
• Add -previousSelectedItemsId which maintains an array of previously
selected item tree identifiers.
• Add -selectionDidChangeToRows:[selectedItem:] which mimics logic of
when row selection is changed so that group items can be rebuilt
wherever.
• Added -adjustSelectionWithRows:selectedItem: which takes a list of
rows that are supposed to be selected + a specific item that is
supposed to be selected and fixes the server list if they are not.
• Added -serverListRowsForItems: which returns an NSIndexSet of an
array of items in the server list. The array can either be a list of
tree identifiers or IRCTreeItem instances. It will check the type to
know which it is.
• -select: will now select the item in current group if the item exists
in the group.
• Added -ignoreNextOutlineViewSelectionChange
• Renamed -ignoreOutlineViewSelectionChanges to
-ignoreOutlineViewSelectionChanges
• Removed -temporarilyDisablePreviousSelectionUpdates — this option is
not needed when -ignoreOutlineViewSelectionChanges is used