Weak link libextension.dylib
Textual swizzles NSExtension which is a symbol of libextension.dylib

Xcode automatically links this library for us since we mention the
symbol in the code, but the library does not exist on OS X Mountain
Lion, which causes a crash when it tries to load.

Weak linking it allows it to fail if it does not exist.