Very large iCloud + theme refactoring
This is a very large commit with many things changed. The most notable
of which are:

• Refactored iCloud support:

- Styles stored on iCloud are no longer copied to a separate folder to
guarantee that they exist in whole. Instead, the folder is checked if
it is 100% downloaded. This greatly simplifies a ton of complexities
when dealing with styles hosted on iCloud.
- Much of code related to iCloud has been rewritten or modified to be
more reliable and easier understand

• Completely rewrote entire sections of the theme controller and
related classes to make them more reliable

• Made a lot of stuff private because plugins