Modernize TDCPreferencesController
Notable changes:
• If a checkbox changes the behavior of another checkbox, then the
enabled state of that checkbox is modified to reflect the change in
behavior so that the user is not caught off guard. KVO is used to
accomplish this so that the defaults value is not modified when
presenting a faked value.
• Added method named -resetWindowFrameToDefault to reset window frame
before saving it when closing so that it can open in correct position
when reopening