Change location in which temporary copy of style is kept
Keeping a cached copy of styles in ~/Library/Caches has proved
problematic because tools that automatically empty these files will
remove style files which leads to a white screen when opening a new
query. They are now kept in the Group Container path for Textual. I am
not aware of any utility that will delete it from there (*fingers
crossed*). This change also makes the folder name have the process ID
as a suffix incase somebody opens more than one copy of Textual, the
second copy does not destroy the cache when quitting.