Refactor dialogs
• Rename all dialogs that begin with “TDChannel” to “TDCChannel”
• Add TDCClientPrototype and TDCChannelPrototype which can be used by a
dialog to declare common properties
• Refactored TDCServerProperties, TDCAddressBookSheet, and
TDCChannelProperties to match modern code base standards such as
nullability, dot-syntax, designated initializers, generics, etc.
• The -config property of TDChannelModifyModesSheet and
TDCHighlightEntrySheet is now strong instead of copy. The designated
initializers for these classes create a copy of input which is then
strong referenced. This is a step towards immutable/mutable
configuration classes. When mutable variants of the configuration
classes are available, the strong reference will be to a mutable copy
which can be tossed around internally.