Refactor IRCChannel + IRCUser
- IRCChannel:
• IRCChannel can only be created by a designated initializer which is
now private. Use IRCWorld instead.
• @syncronized calls for the member list have been removed. No need to
lock the array if we are guaranteeing access on a serial queue.
• Rewrote much of the code related to inserting/removing/modifying
members to be more efficient and/or reduce unnecessary calls.
• Made many features of the class private
• Added property (autoJoin) for modifying the configuration without
accessing it directly.
• -setName: no longer does anything if the channel isn’t a private
message
• -updateConfig: performs validation on new configuration: checks to
make sure the type is the same, the name is the same, and the unique
identifier is the same.

- IRCChannelConfig:
• IRCChannelConfig is now immutable with a mutable variant named
IRCChannelConfig
• Added method named -uniqueCopy which creates a unique copy of the
configuration for purposes such as duplication

- IRCUser:
• IRCUser is now immutable with a mutable variant named IRCUserMutable
• Changing the away status of a user is a common practice. To reduce
overhead of having to create an immutable user just to modify it,
IRCUser now has the methods: -markAsAway and -markAsReturned — these
methods modify the property even if the user is immutable
• Creating a copy of IRCUser as immutable and mutable will pass the
user’s weight so it does not need to be recalculated. The now defunct
-migrate: method did not do this.
• The code for IRCUserNicknameColorStyleGenerator has been operated
into its own file to make it easier to manage.

————

All other changes are related to supporting those mentioned above