Add an 'add' flag to -[IRCWorld createChannelWithConfig:…]
The ‘add’ flag controls whether the method invokes -addChannel: on the
client when creating the channel.

During startup, calling -addChannel: dozens of times is inefficient
because each call requires the client to mutate its configuration and
set it again.

Now an array of channels is created when creating the client so that
the channel list has to only be set once.