Add -extractUserPrefixFromChannelNamed: to IRCISupportInfo
When sending a message to a set of users with a specific mode, a
channel name is prefixed with that mode. For example, the +#channel is
the target of a message directed at voiced users of #channel. This
method extracts the mode from a channel name if present without
confusing channel name characters for modes.