Refactor IRCClient (incomplete)
This commit rewrites large portions of IRCClient

This commit isn't complete either. The refactoring of many commands has
not finished which means they are missing. They will return in later
commits.

The amount that has changed in this commit is very overwhelming and
trying to describe every single change would take a long time.

Key changes:
• All public facing properties (IRCClient.h) are now readonly
• IRCClient can only be initialized through its designated initializer
now, which is private. IRCWorld should be used exclusively for creating
clients.
• -updateConfig: now checks whether the unique identifier for the
configuration being imported matches that of the current configuration.
If they do not match, the new configuration is rejected.
• IRCClientConfig is now immutable which means to make any changes, a
mutable copy would have to be created, modified, then set. There is
overhead in doing this which means values that change frequently (such
as the last server-time value) isn't saved on each change. Instead, the
configuration is written out periodically. When IRCWorld calls
-configurationDictionary on IRCClient to perform a periodic save, the
helper method -updateStoredConfiguration is invoked. This helper method
mutates the config and writes out any pending changes before giving
IRCWorld the most up-to-date values.
• IRCClient now uses -disconnectCallback to properly deconstruct itself
when application is terminating. Previously,
-prepareForApplicationTermination called -quit on the client without
giving it time to disconnect gracefully.
• -cacheHighlightInChannel:withLogLine: will now only add new entry to
the open highlight list sheet if that sheet belongs to the client the
highlight was cached on. This fixes a highlight in client 1 being shown
to user while sheet for client 2 is visible.
• Nothing in IRCClient touches the storage array for the channel list.
Previously, @syncronized was called on the mutable channel list array
maintained by the client whenever a method needed access to the channel
list. Everything now uses the self.channelList proxy so that the list
worked with is immutable (and thus safe).
• Persistent values are now cached using dispatch_once instead of
application-wide shared cache.
• The method for posting text notifications has been merged into the
method for posting event notifications. There is no reason to have two
methods, with duplicate logic.
• All methods that are require a value (non-nil) now throw an exception
for nil value instead of silently failing.
• Methods that require a specific state now check that (e.g.
isConnected, isLoggedIn, etc.) to prevent a plugin from invoking a
method without knowledge of that method's internals, and that method in
turn doing something its not supposed to.
• All methods related to printing to the WebView have been recreated
(new naming)
• Fixed the batch queue not dumping its contents when the client
disconnects
• The list of pending capacities during CAP negotiation is now
maintained in an array instead of integer. This eliminates
-sendNextCapacity having to do some silly tricks to find which CAP to
send next.
• Exposed dozen+ helper methods in IRCClient.h for interacting with IRC
(sending modes, requesting channel topic, etc.) — abstracts away a lot
of nitty gritty for plugins and rest of app.