Add -compileListOfModeChangesForModeSymbol:… to IRCClient
Can take a string and split it up, or take an array of strings. Returns
an array of strings which can be passed to MDOE command to change users
without exceeding mode limit.

For example: Given mode symbol -o and user1 user2 user3 user4 and
maximum count of 3 per-entry, the result of this method will look like:

[
-ooo user1 user2 user3
-o user4
]