Refactored TDCFileTransferDialogTransferController
• All properties are now private and can only be modified through
itself or by one of methods exposed
• Fixed resume not working because -sendTransferResumeAcceptToClient
was calling -sendFileResume: instead of -sendFileResumeAccept:
• -networkInterfaceMatchingAddress now returns the IP address and
nothing else. GCDAsyncSocket is able to go from IP address ➜ interface
on its own.
• Error handling is replaced by -closeWithLocalizedError:… which
formats the localized error message before passing it to the table
cell. This new error handler presents the socket error to the user
instead of logging it to console.
• Modified a few methods to declare a in-scope value so that we are
only setting properties if an operation succeeds
• Dot-syntax