Modernize the system information plugin
• The declaration for TPI_SP_SysInfo is now private (self contained)
• -systemDiskspaceInformation has been rewritten to use NSFileManager
to obtain volume information instead of lower level file system C API
• -systemInformation now builds the output using NSMutableString
instead of appending to an NSString. Additionally, unless a component
is enabled, it no longer requests the information for said component.
• -webKitFrameworkMemoryUsage may return a nil value so its important
to check this fact before asking the client to print it, because the
client now throws an exception for nil values
• Dot-syntax, generics, nullability, etc.