Workaround for white flashes
On styles with a dark background, a white flash occurs because there is
a very small delay between the view being created and the background
process laying out its contents. To work around this,
TVCMainWindowChannelView has been retooled so that its subviews always
show a solid overlay view that is equal to
themeSettings().underlyingWindowColor. This overlay view is presented
until the view has finished loading.

To know accurately when the view has finished loading, a JavaScript
animation frame is used to call the bridged function
app.finishedLayingOutView(). This function then sets the loading flag
to NO on the view, which TVCMainWindowChannelView observes.

Notable changes:

Textual.viewBodyDidLoadInt() is now invoked by body.onload instead of
Textual.viewBodyDidLoad() — this internal function will then call
Textual.viewBodyDidLoad() when the animation frame completes

Textual.viewFinishedLoadingInt() has been modified to cancel the
animation frame for views that are not visible so that the view is
considered "loaded" when the user clicks it.

WKWebView -updateLayer is no longer swizzled

TVCLogView now has the property -layingOutView which
app.finishedLayingOutView() modifies and which TVCMainWindowChannelView
observes.