Modify encryption logic
• Modify method for performing blocks in relations to an account name
- Returns nullable channel value if depending on whether the a new
flag, createWindowIfMissing, is defined. This new flag, which is NO by
default, determines whether a query should be created if one for the
nickname does not already exist on the client. Delegate logic has been
modified to NOT create a new window in most cases.
• Removed message queue. It has been replaced by a delegate callback in
the Encryption Kit framework which can be used to conditionally ignore
messages that are being decoded, thus causing the decode callback block
never being invoked.