Lay foundation for Buddy List window
The Buddy List, which will be used to track the availability of
friends, will soon be available using Control 7 on the keyboard.

It will be restricted to OS X Yosemite and later to make user interface
easier to maintain.

It supports dropping a query from the server list or a user from the
user list onto the window to add an entry.