Significant overhaul to user management
How Textual manages users when connected to a network has changed.

The previous design was for each channel to have unique copy of a user.
This included their nickname, hostmask, modes specific to the channel,
their away status, etc.

Properties such as the user's nickname, hostmask, and away status do
not need to be duplicated.

IRCClient now allocates a single instance of IRCUser to store
properties that are not channel specific.

IRCChannel then stores copies of IRCChannelUser. IRCChannelUser, which
is linked to its IRCUser parent, stores channel specific properties.

---

IRCUser instances may exist regardless of whether a user is within a
channel or not.
If a private message is open with somebody, then an instance of IRCUser
will exist.
An instance of IRCUser will be destroyed after five minutes once it is
no longer in use by a channel or private message.

Destroying it using a timer, rather than immediately, lays ground work
for features that can be activated by recent activity of a user.

All users are destroyed immediately when disconnecting. They are not
maintained between connects.

---

Storing a single instance per-user makes searching for users much
faster. Since we are able to query users by nickname (dictionary key),
there is no need to perform a binary search on the channel.

IRCUser maintains a reference to which IRCChannelUser instance is
assigned to which IRCChannel which means we can quickly lookup IRCUser
then ask that which user is assigned to a specific channel.

---

IRCUserPersistentStore:

- This class is used by IRCUser to store persistent values. Persistent
values are any value that must remain constant between multiple copies
of the same IRCUser instance. When creating a mutable copy of an
IRCUser to change one value, the persistent store for the read-only
version of the user and the mutable version of the user are the same.

---

IRCUserRelations:

- This class is used by IRCUser to map the relationships between
IRCChannelUser and IRCChannel.
- This class is a persistent value is stored in IRCUserPersistentStore