Fix theme validation logic
-performValidationForTheme:… had code built in to favor themes hosted
on iCloud, even when iCloud support was disabled. This created problems
because -themeExists: always returned YES for themes hosted on iCloud
because the logic to say otherwise was not compiled in.

This change adds build conditions around the validation logic so that
iCloud themes are only favored when iCloud support is enabled.