Modify socket behavior
Some users, under some unreplaceable circumstances receive unexpected
EOF on socket after a few minutes. Best efforts have found me unable to
understand why its happening.

This commit changes behavior of app when EOF is received. When it is
received, the property EOFReceived is set on the socket. The pong timer
then forces timeout on the user if the timeout interval is exceeded.

We only shut down on EOF now if we are expecting it. When
isDisconnecting or isQuitting flags are set.