Clean up channel queue code
I have made a hundred modification to this code in the past 48 hours.
Now that I have taken the time to take a close look at it and profile
it, this is the solution that I am happy with.

Current queue design: IRCChannel has two queues.

The first queue (#1) is used exclusively for accessing the internal
mutable array used for the member list. This queue is serial.

The second queue (#2) is used exclusively for performing changes to
members (adding, removing, modifying) in the background. This queue is
serial.

Two queues exist so that we can do work in the background (by queue #2)
does not deadlock queue #1, which can be access from multiple threads.

This commit also gets rid of the member list sorted by nickname length.
TVCLogRenderer is smart enough at this point that it doesn’t need to
scan for nicknames shortest to longest.