Fix incorrect variable passed to plugins
Because newMessage was being passed, any modifications made to the
string by other plugins were completely ignored. We need to pass
returnValue which is the store for the current string, modified by the
last plugin.

This introduced the problem that the Wiki-Style Link Parser plugin
modified the input making it a URL. Then the Smiley Converter plugin
was called with original input, instead of that of the Wiki-Style Link
Parser.