Add new type of IRCChannel: utility
Utility channels act like private messages but are treated differently.

They have nothing to do with IRC. They act as a window to store
messages related to the same event.

• IRCClient uses a utility channel to show server traffic
• The chat filter add-on uses it as the window actions are logged

Utility channels are volatile. They are never saved and never log.

They only exist as long as they are open.
They have no understanding of the events of the client itself.

The user controls menu is completely unavailable in a utility channel.