Modify behavior for loading backlog
Messages reloaded from previous session or after reloading style always
appear in #historic_messages now.

An isreload attribute is now set on #historic_messages with value of
true or false. When the style is reloaded, this value is set to true.
If the value of true is used, a different opacity is used. That way
messages reloaded by changing style do not appear grayed out.