Modify -findChannelOrCreate: logic
When we create a channel through this method, then queue a save on the
world controller. A save is queued, rather than occurring immediately,
so that we do not thrash world controller when multiple channels will
be created.

For example, if we connect to a ZNC bouncer for first time, and that
joins us in a dozen channels for the first time.