Auto scrolling refactoring
• Removed some properties that are no longer used:
TextualScroller.scrollLastPosition3
TextualScroller.scrollHeightTimerActive

• Renamed some variables to better describe what they store:
TextualScroller.scrollHeightCurrentValue ➜
TextualScroller.currentViewHeightValue
TextualScroller.scrollHeightPreviousValue ➜
TextualScroller.previousViewHeightValue

• Added flag TextualScroller.adjustScrollerPosition. When this flag is
set to true, TextualScroller.adjustScrollerPositionInt() is then
invoked when the height of the view changes. This function takes the
difference in height and scrolls to the view based on that, so if
content is added or removed at top of view, the content stays where the
user has scrolled to.

• Renamed TextualScroller.performAutoScrollInt() to
TextualScroller.viewHeightChanged()

• TextualScroller.viewHeightChanged() now records the current height
and the previous height at all times, regardless of whether it returns
true.

• Set defaults key to disable animated scrolling