Refactor member list
Reverted to modifying the member list, including sorting it, on the
same thread. The main thread.

Trying to optimize it by making it perform work in the background has
been a complete nightmare. I have committed fix, to fix the fix, that
fix the fix, that fix the fix, etc. for the past two months+