Save the value of nickname, always
If a query is created for a user that does not already have an instance
of IRCUser, then IRCUser uses the spelling of the nickname that was
created when performing the “/query” command.

This commit modifies the nickname property of IRCUser to reflect that
which the server gives us so that we always have the correct spelling.