Modify logic for /ignore command
Ask for server-wide user instance, instead of the user in the target
channel. This fixes the /ignore command not working if user isn’t in
the channel, or if something other than a channel is selected. The
target channel was never accessed by this command, other than to look
up the user, which means it is not hard to ditch the old logic.