Add TLOSpokenNotification
Because spoken notifications may be delayed by seconds up to minutes
(depending on whether something else is speaking), then it’s better not
to format the notification until it’s time to actually speak it.

TLOSpokenNotification acts as a container for information about a
notification to speak. TLOSpeechSynthesizer has been modified to accept
this object, in addition to regular strings, to make it as easy as
possible to use.

TLOSpeechSynthesizer will automatically call out to IRCClient to format
the notification if the next object in the queue is an
TLOSpokenNotification

Because TLOSpeechSynthesizer and TLOSpokenNotification is aware of what
client created the notification, this allows us to clear the queue of a
specific client as well, such as when deleting the client.