Add code-level support for multiple servers per-connection
This does not have a user interface yet. Trying to modify connection
through Server Properties will throw an exception.

- Overview -

Each server is assigned a server address, port, whether TLS is enabled,
and a password. An unlimited* number can be stored in IRCClientConfig
in the -serverList property.

- Keychain Management -

IRCClient is fully aware of the -serverList property and automatically
writes and destroys keychain items associated with each serve when they
are added or removed.

- Priority -

Each time a new connection is opened, the next server in -serverList is
connected to, starting at index 0 (the first), until it reaches the end
of the array and cycles back.

- Backwards Compatibility -

IRCClientConfig writes out the server address and other properties of
the primary server in -serverList anytime it exports its dictionary.
This is done to allow older versions of Textual to reach the
configuration file without crashing.

Each time IRCClientConfig imports a dictionary, it checks whether the
key migratedToServerListV1Layout is present. If it is not, then it
attempts migration.

If a -serverList array already exists in the dictionary, migration
automatically fails. This is done so that we don’t read the backwards
compatible server information saved in the dictionary for older
versions of Textual.

Server passwords are not backwards compatible. Once the password is
moved from the old keychain item to the server specific one, the old
keychain item is destroyed.