Notification improvements
• Add section to Channel Properties for editing per-channel
notifications. Notification options with a checkbox used mix state to
determine whether the option is on, off, or uses the global default.
• Renamed TDCPreferencesSoundWrapper class to
TDCPreferencesNotificationConfiguration
* Sound options now have a “Default Sound” option which sets the
current -alertSound to nil
• The “None” sound option has been renamed to “No Sound”
• TVCNotificationConfigurationViewController now monitors all keys of
the active alert using KVO so that it can reload the user interface if
the active alert changes remotely. There is no case in which this
happens, yet, but this allows flexibility for the future.