Add three additional sizes for server list
TVCServerListYosemiteUserInterface is now responsible for returning all
values that may pertain to sizes.

TVCServerListYosemiteUserInterface, when initialized, initializes
TVCServerListYosemiteScaledUserInterface or one of its three other
subclasses. Any call to  asking for a constraint then results in it
querying the TVCServerListYosemiteScaledUserInterface object that it
initialized.

When TVCServerListYosemiteScaledUserInterface is initialized, it
records whether the server list is retina. It then returns constraints
based on that fact. Because these classes are reconstructed whenever
Textual changes windows, it will always have correct retina status
without having to recalculate.