Move location in which defaults are registered
The XPC service can host multiple connections. For each connection that
is hosted, a new instance of RCMProcessMain is created. It is
inappropriate to register preferences here.

There is ever only one instance of RCMProcessDelegate.
RCMProcessDelegate is responsible for creating the multiple instances
of RCMProcessMain. It therefore makes much more sense to register
preferences in RCMProcessDelegate, as it will only ever occur once.