Fix incorrect "Remind Me Later" logic
Value should be true, not false, in condition. If user clicked remind me later, than the value should be true, at which time we check last time dialog was checked. We want to show the dialog if user never clicked button to remind them, so this corrected logic makes sense.

We also need to unset key.