Simplify row height calculation
The text field in the table already clips past the default row height when it wraps because it does not have a bottom constraint. We can take advantage of this by taking the frame of the text field and using its height as the height of the row.

No need to perform complex math to determine its height any longer.