Significant TVCLogRenderer rewrite
Instead of using bit shifting to store attributes when renderring, we use an NSAttributedString which is already much better designed for this type of this purpose.

Using NSAttributedString's enumerating gives us better control over the output. We can make it so fragements are only closed when they are supposed to in the HTML. This reduces unncessary output.

formattedMessageFragment.mustache has been modified to support the new fragment closures, but the renderer sets old template tokens for backwards compatiablity, if a style overrides formattedMessageFragment.mustache.