Modify TVCWK1AutoScroller appearance
Scrollbar for WebKit has long had issue with not appearing at all. It would just be white. Spent well over a day reverse engineering WebKit, NScrollView, and other components to resolve this issue to no avail. The scrollbar appears correct in the view hierarchy when inspected in Xcode, but any effort to cause a redraw or layout results in nothing.

This bug was introduced when Textual switched to using -addSubview: with an NSSplitView to show multiple channels at once. Prior, an NSBox was used with its content view set to the WebView.

An effort was made to use an NSBox in the NSSplitView with same outcome.

This change forces scrollbar to always have an "overlay" style which means they only appear when 1) scrolled 2) mouse is near scroll area