Modify how templates are loaded
Textual has long had a "Line Types" folder which contained one template per line type. Each template used a partial to either import newMessagePostedWithSender.mustache or newMessagePostedWithoutSender.mustache. This works fine, mostly, except for when a style tries to change one of those two templates (ewMessagePostedWith*Sender).

Because of the nature of the template loader, the entire Line Types folder would have to be copied to a style to change one of those two templates.

This patch workaround this limitation in the template loader.

Instead of using partials, a property list with the mappings is used. First we try to load the template from the Line Types folder, which doesn't exist by default. If that does not load, we use the property list to map the line type to the appropriate file.

Now the only type the Line Types folder needs to be touched is if a style has a very specific reason to modify a single line type.