Substantial JavaScript refactoring
Functions that are private to Textual and sould never be touched by a style are now assigned to the _Textual object.

Same applies to other objects. There is now _MessageBuffer and _Textual

The visibility of each function is also declared in source, at the end of each declaration, to aid in easier understanding of their scope.