Modify scroller
The scroller is now designed to support more than binding to document.body

When the document loads, _TextualScroller.bindToBestElement() is called. This function will check which element is best to bind to depending on whether overflow-y is set to "none" or not.

A custom style can also call TextualScroller.bindToElement() to bind to the element of its choice.