Modify document body
On <html> set a "visible" and "selected" attribute to allow easier customization through the use of CSS selectors.

For example, pointer-events is now set in baseLayout.css instead of having a JavaScript function to set it.

The topic bar is also hidden now using baseLayout.css

This commit also changes the CSS of default styles to only account for the topic bar height in max-height when selected == true